add to fav button and function, remove fav button and function + list of favourites working

+++ routes/web.php
@@ -2,6 +2,9 @@
 
 use App\Http\Controllers\PhotoController;
 use Illuminate\Support\Facades\Route;
+use Illuminate\Support\Facades\Auth;
+
+
 
 /*
 |--------------------------------------------------------------------------
@@ -18,17 +21,28 @@ Route::get('/', function () {
     return view('home');
 }); 
 
-/* Auth::routes();
- */
-Route::get('/home', [PhotoController::class, 'index']);
-Route::get('/', [App\Http\Controllers\PhotoController::class, 'index'])->name('home');
-Route::delete('/delete/{id}', [PhotoController::class, 'destroy'])->name('remove');
-Route::get('/add', [PhotoController::class, 'create'])->name('newPhoto');
+Auth::routes();
+
+//Routes Create
+Route::get('/add', [PhotoController::class, 'create'])->name('newPhoto')->middleware('isAdmin', 'auth');
 Route::post('/', [PhotoController::class, 'store'])->name('storePhoto');
 
-// Routes edit
-Route::get('/edit/{id}', [PhotoController::class, 'edit'])->name('edit');
+//Route Read
+Route::get('/home', [PhotoController::class, 'index']);
+Route::get('/', [App\Http\Controllers\PhotoController::class, 'index'])->name('home');
+
+// Routes Edit/Update
+Route::get('/edit/{id}', [PhotoController::class, 'edit'])->name('edit')->middleware('isAdmin', 'auth');
 Route::patch('/photo/{id}', [PhotoController::class, 'update'])->name('update');
 
-// Routes show
-Route::get('/detailsPhoto/{id}', [PhotoController::class, 'show'])->name('detail');
+// Route Delete
+Route::delete('/delete/{id}', [PhotoController::class, 'destroy'])->middleware('isAdmin', 'auth')->name('remove');
+
+// Routes Details
+Route::get('/detail/{id}', [PhotoController::class, 'detail'])->name('details');
+
+// Routes Favourites
+Route::get('/favourites', [PhotoController::class, 'listFavourites'])->middleware('auth')->name('favourites');
+Route::get('/addToFav/{id}', [PhotoController::class, 'addToFav'])->middleware('auth')->name('addFav');
+Route::get('/removeFav/{id}', [PhotoController::class, 'removeFromFav'])->middleware('auth')->name('removeFav');
+
